Fix broken links (#52)
* Fix anchor in the "Changing Multiple Commit Messages" link

In
https://github.com/progit/progit2-tr/pull/51/commits/357db9de15cf22d8e1c4945adfdf76788e3059a6#diff-ef001a8fe03c386785e0861e40f85c0db419ea4bf1227a992e163b52f475f82dL450-R449
the anchor was changed from `_changing_multiple` to `_change_multiple`,
breaking the link. Let's change it back to unbreak it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

* Fix the anchors of the "Pushing [...]" links

In
https://github.com/progit/progit2-tr/pull/51/commits/e810e04e74fc854fe1b352764ca731001cf6a3b6,
these three anchors were changed from starting with a lower-case `p` to
an upper-case `P`, breaking the links in the process. Let's downcase
them again to unbreak them.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

---------

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>